+++ src/fscript_ext/fscript_endian.c
@@ -66,6 +66,20 @@ static ret_t func_ntohf(fscript_t* fscript, fscript_args_t* args, value_t* resul
   return RET_OK;
 }
 
+static ret_t func_htonll(fscript_t* fscript, fscript_args_t* args, value_t* result) {
+  FSCRIPT_FUNC_CHECK(args->size == 1, RET_BAD_PARAMS);
+  value_set_int64(result, int64_to_big_endian(value_int64(args->args)));
+
+  return RET_OK;
+}
+
+static ret_t func_ntohll(fscript_t* fscript, fscript_args_t* args, value_t* result) {
+  FSCRIPT_FUNC_CHECK(args->size == 1, RET_BAD_PARAMS);
+  value_set_int64(result, int64_from_big_endian(value_int64(args->args)));
+
+  return RET_OK;
+}
+
 ret_t fscript_endian_register(void) {
   ENSURE(fscript_register_func("is_little", func_is_little_endian) == RET_OK);
 
@@ -78,6 +92,8 @@ ret_t fscript_endian_register(void) {
   ENSURE(fscript_register_func("htonf", func_htonf) == RET_OK);
   ENSURE(fscript_register_func("ntohf", func_ntohf) == RET_OK);
 
+  ENSURE(fscript_register_func("htonll", func_htonll) == RET_OK);
+  ENSURE(fscript_register_func("ntohll", func_ntohll) == RET_OK);
 
   return RET_OK;
 }